requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
cryptography==44.0.1
Django==5.1.6
django-allauth==65.4.1
django-appconf==1.1.0
django-crispy-forms==2.3
django-cryptography==1.1
idna==3.10
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
